fixed underdrawing
I currently ADD to a pixel, because I am doing everything in one loop. This is causing OVERDRAWING.
Once I implement an acceleration structure which only renders to a AABB of the triangle and runs the loop for it then this issue will be fixed because I'll SET the colour of that pixel.